requirements.txt
openai==1.84.0
python-dotenv==1.1.0
Requests==2.32.3
selenium==4.33.0
seleniumbase==4.39.2
